# packages/spec/openapi.yaml
openapi: 3.1.0
info:
  title: Bark & Meow — API
  version: 1.0.0
  description: |-
    Contrato de apps/api: lo consumen el portal del dueño (mi-mascota), la app, la web de la clínica, la web del veterinario, el panel del operador y el software de gestión.

    Dos principios que cualquier implementación debe respetar:

    1. **El servidor no descifra.** Todo lo que lleva contenido clínico viaja como bytes opacos en base64.
    2. **Localizar no es abrir.** La consulta por identificador responde si existe ficha y nada más, con el mismo tamaño y el mismo tiempo en ambos casos.

    La consulta entre nodos (`/federation/v1/lookup`) está por implementar: solo podrá decir si existe ficha y ofrecer un enlace de aviso, con peticiones y respuestas firmadas con Ed25519.

    Generado desde packages/schema/src/api. No se edita a mano.
  license:
    name: CC BY 4.0
    url: https://creativecommons.org/licenses/by/4.0/
servers:
  - url: https://barkandmeow.app/api
    description: Producción
  - url: http://127.0.0.1:4601
    description: Desarrollo
paths:
...
  /chip/v1/notify:
    post:
      tags:
        - "Nivel 0: localizar"
      summary: Aviso sellado para el dueño con el token que dio el nivel 0. Responde igual con token real o con señuelo, así que no confirma si el chip existe.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $schema: https://json-schema.org/draft/2020-12/schema
              type: object
              properties:
                aviso:
                  type: string
                  pattern: ^[A-Za-z0-9_-]{40,96}$
                sellado:
                  type: string
                  pattern: ^[A-Za-z0-9+/]+={0,2}$
              required:
                - aviso
                - sellado
      responses:
        "202":
          description: Correcto
          content:
            application/json:
              schema:
                $schema: https://json-schema.org/draft/2020-12/schema
                type: object
                properties:
                  recibido:
                    type: boolean
                    const: true
                required:
                  - recibido
                additionalProperties: false
        "400":
          description: Error 400
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-acceso: web-vet
...
  /e/v1/{id}:
    get:
      tags:
        - Web del veterinario
      summary: "Nivel 1: el resumen de emergencia de una placa, sellado. La clave va en el fragmento de la URL y nunca llega aquí."
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Correcto
          content:
            application/json:
              schema:
                $schema: https://json-schema.org/draft/2020-12/schema
                type: object
                properties:
                  sobre:
                    type: string
                    pattern: ^[A-Za-z0-9+/]+={0,2}$
                  version:
                    type: integer
                    minimum: -9007199254740991
                    maximum: 9007199254740991
                  perfil:
                    anyOf:
                      - type: object
                        properties:
                          nombre:
                            type: string
                          bio:
                            type: string
                          telefonos:
                            type: array
                            items:
                              type: object
                              properties:
                                etiqueta:
                                  type: string
                                numero:
                                  type: string
                              required:
                                - etiqueta
                                - numero
                              additionalProperties: false
                          foto:
                            type:
                              - string
                              - "null"
                        required:
                          - nombre
                          - bio
                          - telefonos
                          - foto
                        additionalProperties: false
                      - type: "null"
                required:
                  - sobre
                  - version
                  - perfil
                additionalProperties: false
        "404":
          description: Error 404
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-acceso: web-vet
  /s/v1/{id}:
    get:
      tags:
        - Web del veterinario
      summary: "Nivel 2: la copia temporal (historial o pasaporte) mientras no caduque, con la clave pública del dueño para sellar la nota."
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Correcto
          content:
            application/json:
              schema:
                $schema: https://json-schema.org/draft/2020-12/schema
                type: object
                properties:
                  sobre:
                    type: string
                    pattern: ^[A-Za-z0-9+/]+={0,2}$
                  caduca:
                    type: string
                    format: date-time
                    pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
                  ownerPubKey:
                    type: string
                    pattern: ^[A-Za-z0-9+/]+={0,2}$
                  perfil:
                    anyOf:
                      - type: object
                        properties:
                          nombre:
                            type: string
                          bio:
                            type: string
                          telefonos:
                            type: array
                            items:
                              type: object
                              properties:
                                etiqueta:
                                  type: string
                                numero:
                                  type: string
                              required:
                                - etiqueta
                                - numero
                              additionalProperties: false
                          foto:
                            type:
                              - string
                              - "null"
                        required:
                          - nombre
                          - bio
                          - telefonos
                          - foto
                        additionalProperties: false
                      - type: "null"
                required:
                  - sobre
                  - caduca
                  - ownerPubKey
                  - perfil
                additionalProperties: false
        "404":
          description: Error 404
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "410":
          description: Error 410
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-acceso: web-vet
  /s/v1/{id}/doc/{docId}:
    get:
      tags:
        - Web del veterinario
      summary: Documento original de un registro, cifrado con la misma clave temporal.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: docId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bytes
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          description: Error 404
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "410":
          description: Error 410
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-acceso: web-vet
  /s/v1/{id}/nota:
    post:
      tags:
        - Web del veterinario
      summary: "Nota de la consulta, sellada para el dueño en el navegador del veterinario: escribe, no lee."
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $schema: https://json-schema.org/draft/2020-12/schema
              type: object
              properties:
                sellado:
                  type: string
                  pattern: ^[A-Za-z0-9+/]+={0,2}$
              required:
                - sellado
      responses:
        "201":
          description: Correcto
          content:
            application/json:
              schema:
                $schema: https://json-schema.org/draft/2020-12/schema
                type: object
                properties:
                  ok:
                    type: boolean
                    const: true
                required:
                  - ok
                additionalProperties: false
        "400":
          description: Error 400
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Error 404
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "410":
          description: Error 410
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-acceso: web-vet
  /perfil/v1/foto/{id}:
    get:
      tags:
        - Web del veterinario
      summary: "Foto del perfil público, solo si el dueño lo ha publicado. El id cambia con cada foto: se cachea como inmutable."
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bytes
          content:
            image/*:
              schema:
                type: string
                format: binary
        "404":
          description: Error 404
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-acceso: web-vet
...
components:
  securitySchemes:
    sesionDueno:
      type: apiKey
      in: cookie
      name: bam_owner
      description: Portal del dueño.
    tokenApp:
      type: http
      scheme: bearer
      description: "App: token de sesión, con la cabecera `x-bm-cliente: app`."
    sesionClinica:
      type: apiKey
      in: cookie
      name: bam_clinic
      description: Identifica a la organización y a la persona. No es una llave a los datos.
    claveApi:
      type: http
      scheme: bearer
      description: "Software de gestión: `bmk_…`."
    tokenOperador:
      type: http
      scheme: bearer
      description: "Panel del operador: OPS_TOKEN."
  schemas:
    Error:
      $schema: https://json-schema.org/draft/2020-12/schema
      type: object
      properties:
        error:
          type: string
        motivo:
          type: string
      required:
        - error
      additionalProperties: {}